requirements.txt
pytesseract
wikipedia
openpyxl
flask
Pillow
opencv-python